src/requirements.txt
Flask==1.0.2
autopep8==1.4
ipython==5.8.0
picamera==1.13
protobuf==3.6.1
pyyaml>=4.2b1
rpi.gpio==0.6.3
schedule==0.5.0
virtualenv==16.0.0
smbus
